requirements.txt
flask==3.1.0
groq==0.22.0
python-dotenv==1.1.0 